chore: moves DataEntryBottomDialog to form module
