+++ app/src/main/java/org/dhis2/usescases/eventsWithoutRegistration/eventCapture/model/EventCompletionButtons.kt
@@ -1,6 +1,6 @@
 package org.dhis2.usescases.eventsWithoutRegistration.eventCapture.model
 
-import org.dhis2.ui.DialogButtonStyle
+import org.dhis2.form.ui.dialog.DialogButtonStyle
 import org.dhis2.utils.customviews.FormBottomDialog
 
 data class EventCompletionButtons(
