+++ form/build.gradle
@@ -47,6 +47,8 @@ android {
     }
 
     buildFeatures {
+        compose true
+
         dataBinding true
     }
 
@@ -66,6 +68,10 @@ android {
 
     kotlinOptions {
         jvmTarget = '1.8'
+    }
+
+    composeOptions {
+        kotlinCompilerExtensionVersion = '1.0.1'
     }
 }
