+++ form/src/main/res/values-es/strings.xml
@@ -48,4 +48,5 @@
     <string name="due_date">Fecha de vencimiento</string>
     <string name="event_date">Fecha de evento</string>
     <string name="mandatory_field">Campo obligatorio</string>
+    <string name="discard_changes">Descartar los cambios</string>
     </resources>
