+++ form/src/main/res/values/strings.xml
@@ -51,4 +51,5 @@
     <string name="mandatory_field">Mandatory field</string>
     <string name="form_latitude_prefix">Lat:</string>
     <string name="form_longitude_prefix">Long:</string>
+    <string name="discard_changes">Discard changes</string>
 </resources>
